CDCN/requirements.txt
Cython==0.29.24
h5py==3.1.0
imageio==2.9.0
imgaug==0.4.0
matplotlib==3.3.4
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
pandas==1.1.5
Pillow==8.3.2
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
sklearn==0.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
torch==1.8.0+cu111
torchsummary==1.5.1
torchvision==0.9.0+cu111
tqdm==4.62.3
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1622748266870/work